requirements.txt
altgraph 
ascii-magic==2.3.0
asgiref==3.8.1
autoflake==2.3.1
certifi==2024.12.14
charset-normalizer==3.4.1
climage==0.2.2
colorama==0.4.6
contourpy==1.3.0
coverage==7.6.12
crispy-bootstrap4==2024.10
cycler==0.12.1
Django==4.2.19
django-crispy-forms==2.3
exceptiongroup==1.2.2
factory_boy==3.3.3
Faker==36.1.1
fonttools==4.55.3
idna==3.10
importlib_resources==6.5.2
inflection==0.5.1
iniconfig==2.0.0
isort==5.13.2
kdtree==0.16
kiwisolver==1.4.7
macholib 
matplotlib==3.9.4
numpy==2.0.2
packaging==24.2
pillow==11.1.0
pluggy==1.5.0
pyflakes==3.2.0
pyparsing==3.2.1
pytest==8.3.4
pytest-django==4.10.0
pytest-factoryboy==2.7.0
python-dateutil==2.9.0.post0
qrcode==8.0
qrcodeT==1.0.4
requests==2.32.3
simple-imshow==1.0.post1
sqlparse==0.5.3
terminal-img==0.0.3
tomli==2.2.1
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
zipp==3.21.0
